rename streamerRepositoryV2 to streamerREpository
