+++ lib/streamerRepository.js
