+++ lib/streamerService.js
@@ -1,6 +1,6 @@
 const Discord = require('discord.js');
 const twitchService = require('./twitchService');
-const { StreamerRepositoryV2 } = require('./streamerRepositoryV2');
+const { StreamerRepositoryV2 } = require('./streamerRepository');
 const { TwitchRepository } = require('./twitchRepository');
 
 function getStreamerEmbded(streamerChannel, user) {
